frontend/backend/requirements.txt
requests
fastapi
uvicorn
Jinja2
httpx
pytest
supabase
python-dotenv
